refactor: Call a page a page
