+++ src/client/pages/ConsentsNewslettersPage.tsx
