--- src/client/pages/NotFound.stories.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Meta } from '@storybook/react';
-
-import { NotFound } from './NotFound';
-
-export default {
-  title: 'Pages/NotFound',
-  component: NotFound,
-} as Meta;
-
-export const Default = () => <NotFound />;
