+++ src/client/pages/UnexpectedErrorPage.tsx
@@ -12,14 +12,13 @@ const link = css`
   display: inline-block;
 `;
 
-export const NotFound = () => (
+export const UnexpectedErrorPage = () => (
   <SignInLayout>
     <PageBox>
-      <PageHeader>Sorry – the page does not exist</PageHeader>
+      <PageHeader>Sorry – an unexpected error occurred</PageHeader>
       <PageBody>
         <PageBodyText>
-          You may have followed an outdated link, or have mistyped a URL. If you
-          believe this to be an error, please{' '}
+          An error occurred, please try again or{' '}
           <Link css={link} href={locations.REPORT_ISSUE}>
             report it
           </Link>
