+++ src/client/routes.tsx
@@ -2,19 +2,19 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { ResetPasswordPage } from '@/client/pages/ResetPasswordPage';
 import { ResetSentPage } from '@/client/pages/ResetSentPage';
-import { Routes } from '@/shared/model/Routes';
-import { NotFound } from '@/client/pages/NotFound';
-import { ChangePasswordPage } from '@/client/pages/ChangePassword';
+import { NotFoundPage } from '@/client/pages/NotFoundPage';
+import { ChangePasswordPage } from '@/client/pages/ChangePasswordPage';
 import { ChangePasswordCompletePage } from '@/client/pages/ChangePasswordCompletePage';
 import { ResendPasswordPage } from '@/client/pages/ResendPasswordPage';
 import { ConsentsDataPage } from '@/client/pages/ConsentsDataPage';
-import { ConsentsCommunicationPage } from '@/client/pages/ConsentsCommunication';
-import { ConsentsNewslettersPage } from '@/client/pages/ConsentsNewsletters';
+import { ConsentsCommunicationPage } from '@/client/pages/ConsentsCommunicationPage';
+import { ConsentsNewslettersPage } from '@/client/pages/ConsentsNewslettersPage';
 import { ConsentsConfirmationPage } from '@/client/pages/ConsentsConfirmationPage';
 import { ResendEmailVerificationPage } from '@/client/pages/ResendEmailVerificationPage';
+import { UnexpectedErrorPage } from '@/client/pages/UnexpectedErrorPage';
+import { ConsentsFollowUpPage } from '@/client/pages/ConsentsFollowUpPage';
 import { ClientState } from '@/shared/model/ClientState';
-import { UnexpectedError } from '@/client/pages/UnexpectedError';
-import { ConsentsFollowUp } from './pages/ConsentsFollowUp';
+import { Routes } from '@/shared/model/Routes';
 
 export type RoutingConfig = {
   clientState: ClientState;
@@ -58,23 +58,23 @@ export const GatewayRoutes = () => (
       exact
       path={`${Routes.CONSENTS}${Routes.CONSENTS_FOLLOW_UP_NEWSLETTERS}`}
     >
-      <ConsentsFollowUp />
+      <ConsentsFollowUpPage />
     </Route>
     <Route
       exact
       path={`${Routes.CONSENTS}${Routes.CONSENTS_FOLLOW_UP_CONSENTS}`}
     >
-      <ConsentsFollowUp />
+      <ConsentsFollowUpPage />
     </Route>
     {/*  ABTEST: followupConsent : END */}
     <Route exact path={Routes.VERIFY_EMAIL}>
       <ResendEmailVerificationPage />
     </Route>
     <Route exact path={Routes.UNEXPECTED_ERROR}>
-      <UnexpectedError />
+      <UnexpectedErrorPage />
     </Route>
     <Route>
-      <NotFound />
+      <NotFoundPage />
     </Route>
   </Switch>
 );
